Support convert big number

+++ NumberToText/NumberToText/IntToWord.h
@@ -14,20 +14,20 @@ public:
 	IntToWord();
 	
 	//Overload Construtor
-	IntToWord(int, string);
+	IntToWord(long long int, string);
 
 	//Destructor
 	~IntToWord();
 
 	//Accessor Functions
-	int getNum() const;
+	long long int getNum() const;
 		// getNum - return number that we want to convert
 
 	string getLang() const;
 		// getLang - return language of converted number
 
 	//Mutator Functions
-	void setNum(int);
+	void setNum(long long int);
 		// setNum - sets number that be convert
 		// @param int - number that be convert
 
@@ -35,24 +35,24 @@ public:
 		// setLang - sets language that converted wanna be
 		// @param string - language that converted wanna be
 
-	string convert_nn(int);
+	string convert_nn(long long int);
 		// convert number to word, a value < 100 to English.
 		// @param int - number that will be convert
 		// @return string - number in word
 
-	string convert_nnn(int);
+	string convert_nnn(long long int);
 		// convert number to word, a value > 100 and < 1000.
 		// @param int - number that will be convert
 		// @return string - number in word
 
-	string convert_number(int);
+	string convert_number(long long int);
 		// convert number to word.
 		// @param int - number that will be convert
 		// @return string - number in word
 
 private:
 	//Member Variables
-	int newNum;
+	long long int newNum;
 	string newLang;
 	static const char* const ones[];
 	static const char* const tens[];
